Directory cleanup
moving files to clean up directory. Moved data files for node analysis
in D3 visualization to separate directory. It was fun, but slow and may
not be useful. Going to take advantage of some R libraries for this
type of analysis if it works out. I’d prefer a “chord” diagram which
might be easier to do in R? not sure.